feat(preprocess): full/binary/continuous variants + memory-safe tokenizer
- build_dataset: add --condition {failonly,full} and --outcome-type {binary,
  continuous}. Output filename now includes both axes:
  ``{method}_{condition}_{outcome_type}.parquet``.
- Pre-tokenization size filter (>50 steps OR env_io >200KB) prevents OOM on
  outlier trajectories (e.g., agent infinite loops with 990MB env_io).
- tokenize_trajectory: refactor from incremental O(N^2) per-step chat-template
  invocations to single-shot O(N): render once, char-position boundaries via
  deterministic accounting + offset_mapping for char→token mapping. Robust
  to <|im_start|>/<|im_end|> substrings embedded in message content.
- Configs and orchestration scripts for vanilla_full_binary and
  vanilla_full_continuous variants.
- New test_outcome_loading.py covering binary/continuous loaders + empirical
  cross-seed variance check on real data.

Verification on Full pool (89/90 groups, 712 rows; 1 outlier skipped):
  binary outcome: 9/89 groups have signal (matches the 9 tasks with ≥1 binary
                  success across seeds)
  continuous:     71/89 groups have signal (variance from partial-pass rates)

+++ grpo/preprocess/make_advantages.py
@@ -24,7 +24,7 @@ class RolloutData:
     response_mask: list[int]
     step_token_ranges: list[tuple[int, int]]
     segments: list[dict[str, Any]]
-    outcome: int
+    outcome: float  # ∈ [0, 1]; binary {0.0, 1.0} or continuous pass_rate
     token_adv: np.ndarray = field(default=None)  # type: ignore[assignment]
 
     def __post_init__(self) -> None:
